fix(Measure CPU): fix Windows problem to finish process

+++ app/measurement/cpu/measure_cpu.py
@@ -38,9 +38,7 @@ class CpuMeasurement(Thread):
         self.tdp = cpu_voltage_file[cpu_voltage_file['PROCESSOR'] == cpu]['TDP'].values[0]
 
     def get_measurents(self):
-        timeout = 0
-        while self.process.is_running() and timeout < 100 and (self.process.status() != psutil.STATUS_ZOMBIE):
-            timeout += 1
+        while self.process.is_running():
             try:
                 num_cores = psutil.cpu_count()
                 children = self.process.children(recursive=True)
@@ -49,25 +47,30 @@ class CpuMeasurement(Thread):
                         self.measure_cpu(child, num_cores)
                 else:
                     self.measure_cpu(self.process, num_cores)
+
+                if self.process.status() == psutil.STATUS_ZOMBIE:
+                    self.process.terminate()
+                    self.process.wait()
             except Exception as e:
                 pass
-        self.process.terminate()
-        self.process.wait()
+
         df_content = {
             "time":self.time_list,
             "cpu_usage": self.cpu_use_list
         }
+
         df_cpu = pd.DataFrame(df_content)
-        df_cpu = df_cpu.iloc[:-1]
+        df_cpu = df_cpu[df_cpu['cpu_usage'] != 0]
         df_cpu.to_csv("app/measurement/test_result/test.csv",index=False,header=True)
         return
 
     def measure_cpu(self, process, num_cores):
         cpu_usage = process.cpu_percent(interval=1)
-        print(f"USO DA CPU: {cpu_usage / num_cores}")
+        cpu_usage_single = cpu_usage / num_cores
+        print(f"USO DA CPU: {cpu_usage_single}")
         self.time_list.append(datetime.now())
-        self.cpu_use_list.append(cpu_usage / num_cores)
-        return cpu_usage / num_cores
+        self.cpu_use_list.append(cpu_usage_single)
+        return cpu_usage_single
 
     def start_measurent(self, process_id):
         self.process = psutil.Process(process_id)
